# lv2: sys_timer_usleep improvements for linux