This is inspired by: https://github.com/RPCS3/rpcs3/pull/4716
This commit causes usleep to only sleep until one quantum before the requested time on linux. There is no change to how usleep works on windows. 


<details>
  <summary>before</summary>

```
Application started
Calculating baseline delay...(0us)
Testing usleep(0)...
    Latency = 0us
Testing usleep(1)...
    Latency = 2us
Testing usleep(10)...
    Latency = 11us
Testing usleep(50)...
    Latency = 50us
Testing usleep(100)...
    Latency = 100us
Testing usleep(200)...
    Latency = 281us
Testing usleep(300)...
    Latency = 385us
Testing usleep(400)...
    Latency = 493us
Testing usleep(500)...
    Latency = 585us
Doing the special tests...
Testing usleep(250)...
    Latency = 274us
Testing usleep(280)...
    Latency = 285us
Testing usleep(290)...
    Latency = 294us
Testing usleep(299)...
    Latency = 302us
Testing usleep(301)...
    Latency = 390us
Testing usleep(310)...
    Latency = 389us
Testing usleep(320)...
    Latency = 389us
Testing usleep(600)...
    Latency = 695us
Application finished
```


</details>

<details>
<summary>after</summary>

```
Application started
Calculating baseline delay...(0us)
Testing usleep(0)...
    Latency = 0us
Testing usleep(1)...
    Latency = 2us
Testing usleep(10)...
    Latency = 11us
Testing usleep(50)...
    Latency = 51us
Testing usleep(100)...
    Latency = 101us
Testing usleep(200)...
    Latency = 201us
Testing usleep(300)...
    Latency = 304us
Testing usleep(400)...
    Latency = 408us
Testing usleep(500)...
    Latency = 510us
Doing the special tests...
Testing usleep(250)...
    Latency = 253us
Testing usleep(280)...
    Latency = 283us
Testing usleep(290)...
    Latency = 293us
Testing usleep(299)...
    Latency = 303us
Testing usleep(301)...
    Latency = 304us
Testing usleep(310)...
    Latency = 314us
Testing usleep(320)...
    Latency = 324us
Testing usleep(600)...
    Latency = 612us
Application finished
```

</details>

This causes us to spend less time sleeping, but we're still able to sleep most of the time on linux. 

On a simple scene PPU utilization looks like this after the change:
![ppuutil1](https://user-images.githubusercontent.com/26352541/59555846-5cb16f80-8f87-11e9-8417-d564c38fe70f.png)


Before, on linux it would look like this: 
![ppuutil2](https://user-images.githubusercontent.com/26352541/59555774-9bdec100-8f85-11e9-8156-24002c04ca10.png)

On windows, it looks like this: 
![ppuutil3](https://user-images.githubusercontent.com/26352541/59555777-af8a2780-8f85-11e9-8ccf-4eede657a66e.png)

Even though its dramatically more accurate now, we are only spending a small amount of time yielding, most of the time we can sleep and not overload low thread cpus. 

Spent awhile looking through my games to see if any benefited, but none seemed to. But there were also no speed regressions, atleast on my 4C8T cpu. 

Finally, I've added a TODO message asking to confirm whether Apple and or any BSD can benefit from having the min quantum set lower. If anyone could test that then we can remove that and those operating systems can benefit as well.