@ruipin 2019-06-16
Did you test any games? Last time I played around with usleep (including seeing what happens when you skip the last quantum) a *ton* of games broke. In particular, some games (I believe Journey was one) do 1us sleeps instead of yielding, so skipping the last quantum just completely broke the game.