@Whatcookie 2019-06-16
@ruipin I tested all the games I have. I tested Nier a bit, and that game spams 1us sleeps as well, and it still works. 

Besides though, this won't change behavior for games requesting sleeps that small anyways, since we still don't sleep for values that small. 

I know that some games can be tolerant of terrible usleep timing (I messed around with it a lot when I was trying to find the speed difference between windows and linux in that game, and you can do literally anything (including intentionally breaking it) and the game will still play fine)