@ruipin 2019-06-16
Journey requests sleeps that vary in size, and can go above the minimum quantum value on linux.

The problem is some games are extremely sensitive, and so are some CPUs. 4 thread CPUs especially are very sensitive. I wouldn't merge such a change without a *lot* of testing.