@Whatcookie 2019-06-16
I'm not sure that this is as unstable as you think it is, we aren't "skipping" the last quantum with this commit, we're making sure that we handle it accurately. 

The windows code is inaccurate for values above 500 us, and the linux code on master is inaccurate for values above 100 us. 

You can see in the results I posted above, while 100us works fine on master, just 200us suddenly causes linux to sleep for 281us. Even such small values can be wildly inaccurate on master.

That being said, while I don't expect any negative effects other than ~2% higher cpu usage seen in the overlay, I haven't found any games that respond positively to the added accuracy. So there's certainly no rush to get this merged.

I'll test this on my computer with hypthreading disabled later and see how bad it is 4T cpus, but testing from people with other usleep sensitive games would be nice. 

If this method is really determined too expensive for 4T cpus (I doubt it), then I have another commit that I can push that I was playing around with earlier that should make usleep about 50% more accurate on linux (without spending any extra time yielding).