@ruipin 2019-06-16
You need to specifically test 2C4T CPUs. I remember the windows hack (where we don't spinlock for the last quantum for accuracy) was put into place *specifically* for those.