@Whatcookie 2019-06-16
Disabled 2 cores of my 4C8T cpu (7700K) as per ruipin's suggestion and retested all my games.

Even in situations with 100% cpu load there didn't seem to be any difference in framerate between master and this pr. I put extra testing into P5 since rui mentioned that many users on windows with 2C4T cpus complained about loss of performance in that game when usleep was first made more accurate. 

That being said, no games see any measurable improvement with 4C8T enabled either, so it'd be great if more people could test this PR.