@Whatcookie 2019-06-18
I've pushed a new commit that scraps the old idea entirely, but this new one is now much improved over master with no drawbacks at all.

The default minimum quantum on linux appears to be 50 microseconds on linux. Now we don't have to spin as much as before for values between 50-100 microseconds. 

There is also an accuracy improvement for values above 100 microseconds, and this is achieved without any extra spinning compared to master. 

100 microseconds is a fairly common value for games to sleep for, and those games see improved performance on my 7700K with this new commit: 

Journey demo before
![journeybefore](https://user-images.githubusercontent.com/26352541/59724966-d5b7fd80-91f9-11e9-8dac-78e4ff206280.png)

Journey demo after
![journeyafter](https://user-images.githubusercontent.com/26352541/59724967-d6509400-91f9-11e9-84a1-c2db6798c5b2.png)





<details>
  <summary>master</summary>

```
Application started
Calculating baseline delay...(0us)
Testing usleep(0)...
    Latency = 0us
Testing usleep(1)...
    Latency = 2us
Testing usleep(10)...
    Latency = 11us
Testing usleep(50)...
    Latency = 50us
Testing usleep(100)...
    Latency = 100us
Testing usleep(200)...
    Latency = 281us
Testing usleep(300)...
    Latency = 385us
Testing usleep(400)...
    Latency = 493us
Testing usleep(500)...
    Latency = 585us
Doing the special tests...
Testing usleep(250)...
    Latency = 274us
Testing usleep(280)...
    Latency = 285us
Testing usleep(290)...
    Latency = 294us
Testing usleep(299)...
    Latency = 302us
Testing usleep(301)...
    Latency = 390us
Testing usleep(310)...
    Latency = 389us
Testing usleep(320)...
    Latency = 389us
Testing usleep(600)...
    Latency = 695us
Application finished
```


</details>


<details>
<summary>this pr</summary>

```
Application started
Calculating baseline delay...(0us)
Testing usleep(0)...
    Latency = 0us
Testing usleep(1)...
    Latency = 2us
Testing usleep(10)...
    Latency = 11us
Testing usleep(50)...
    Latency = 51us
Testing usleep(100)...
    Latency = 102us
Testing usleep(200)...
    Latency = 223us
Testing usleep(300)...
    Latency = 329us
Testing usleep(400)...
    Latency = 441us
Testing usleep(500)...
    Latency = 547us
Doing the special tests...
Testing usleep(250)...
    Latency = 273us
Testing usleep(280)...
    Latency = 284us
Testing usleep(290)...
    Latency = 293us
Testing usleep(299)...
    Latency = 302us
Testing usleep(301)...
    Latency = 332us
Testing usleep(310)...
    Latency = 331us
Testing usleep(320)...
    Latency = 334us
Testing usleep(600)...
    Latency = 650us
Application finished

```

</details>

I believe this is ready for review now.