@kd-11 2019-06-19
Wouldn't shifting the load to OS yield() cause problems for low-end CPUs? A 7700K isn't a good enough test for this imo, we had issues with this code in the past when tested on intel's 4 thread CPUs (older gen i3s and i5s) so I think we need more user testing here.
As for the code itself, its good.