@Whatcookie 2019-06-19
The latest commit doesn't end up yielding any more than master does, the accuracy improvements for larger values were achieved just by sleeping better. 

Two games from my testing see time spent yielding, Persona 5 has about 1-2% less PPU usage with this commit, and Journey has 8% less PPU usage.