@Whatcookie 2019-06-19
Did some comparisons with 2C4T enabled in bios with the new commit: 
### Journey master
![journey2C4Tmaster](https://user-images.githubusercontent.com/26352541/59802978-2a21b280-92b8-11e9-8c8b-0471a10eac6e.png)
### Journey PR
![journey2C4TPR](https://user-images.githubusercontent.com/26352541/59802981-2beb7600-92b8-11e9-82a6-745b01d332de.png)


### Persona 5 master
![P52C4Tmaster](https://user-images.githubusercontent.com/26352541/59803072-6f45e480-92b8-11e9-856a-cad1aa3f88f8.png)



### Persona 5 PR
![P52C4TPR](https://user-images.githubusercontent.com/26352541/59803074-70771180-92b8-11e9-8e6f-0b8c84a1964e.png)
### Ninja gaiden master
![NGSmaster](https://user-images.githubusercontent.com/26352541/59803100-8389e180-92b8-11e9-8587-c028fe125348.png)

### Ninja gaiden PR
![NGSPR](https://user-images.githubusercontent.com/26352541/59803105-84227800-92b8-11e9-9dc0-61a6fe5c13a6.png)

For games that sleep with values between 50-100 us there is a large reduction in time spent yielding, but there is a small increase in time spent yielding for other games that I could not see when I was testing on 4C8T.